auth 🔐 + major refactor
