--- app/login/page.tsx
@@ -1,11 +1,0 @@
-'use client'
-
-import Login from "@/components/Login";
-
-export default function App() {
-  return (
-    <main>
-        <Login />
-    </main>
-  );
-}
