--- app/page.tsx
@@ -1,11 +1,0 @@
-'use client'
-
-import Landing from "@/components/Landing";
-
-export default function App() {
-  return (
-    <main>
-      <Landing />
-    </main>
-  );
-}
