--- app/signup/page.tsx
@@ -1,11 +1,0 @@
-'use client'
-
-import SignUp from "@/components/SignUp";
-
-export default function App() {
-  return (
-    <main>
-        <SignUp />
-    </main>
-  );
-}
