+++ package.json
@@ -13,6 +13,7 @@
     "bcrypt": "^5.1.1",
     "firebase": "^10.12.3",
     "next": "14.2.4",
+    "next-auth": "^4.24.7",
     "react": "^18",
     "react-countup": "^6.5.3",
     "react-dom": "^18",
