+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,0 +1,6 @@
+import { authOptions } from '@/src/server/auth';
+import NextAuth from 'next-auth';
+
+// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
+const handler = NextAuth(authOptions);
+export { handler as GET, handler as POST };
