+++ src/components/Navbar.tsx
@@ -1,13 +1,16 @@
+'use client';
+
 import React, { useState, useEffect } from 'react';
 import Link from 'next/link';
 
-const Navbar = () => {
+export default function Navbar() {
   const [isScrollingDown, setIsScrollingDown] = useState(false);
   const [lastScrollTop, setLastScrollTop] = useState(0);
 
   useEffect(() => {
     const handleScroll = () => {
-      const scrollTop = window.pageYOffset || document.documentElement.scrollTop;
+      const scrollTop =
+        window.pageYOffset || document.documentElement.scrollTop;
       if (scrollTop > lastScrollTop) {
         setIsScrollingDown(true);
       } else {
@@ -21,7 +24,11 @@ const Navbar = () => {
   }, [lastScrollTop]);
 
   return (
-    <nav className={`bg-transparent fixed top-0 z-50 w-full transition-transform duration-300 ${isScrollingDown ? '-translate-y-full' : 'translate-y-0'}`}>
+    <nav
+      className={`bg-transparent fixed top-0 z-50 w-full transition-transform duration-300 ${
+        isScrollingDown ? '-translate-y-full' : 'translate-y-0'
+      }`}
+    >
       <div className="w-100%">
         <div className="flex justify-between h-16">
           <div className="flex items-center sm:px-5 sm:pr-20">
@@ -30,21 +37,36 @@ const Navbar = () => {
             </Link>
           </div>
           <div className="flex items-center justify-end flex-1 sm:space-x-10 md:space-x-20 sm:pr-10 md:pr-20 xl:pr-40">
-            <Link href="/home" className="text-gray-700 hover:text-gray-900 text-xl">
+            <Link
+              href="/home"
+              className="text-gray-700 hover:text-gray-900 text-xl"
+            >
               Home
             </Link>
-            <Link href="/projects" className="text-gray-700 hover:text-gray-900 text-xl">
+            <Link
+              href="/projects"
+              className="text-gray-700 hover:text-gray-900 text-xl"
+            >
               Projects
             </Link>
-            <Link href="/contact" className="text-gray-700 hover:text-gray-900 text-xl">
+            <Link
+              href="/contact"
+              className="text-gray-700 hover:text-gray-900 text-xl"
+            >
               Contact
             </Link>
           </div>
           <div className="flex items-center space-x-10 pr-5">
-            <Link href="/login" className="text-gray-700 border-2 border-gray-600 rounded-xl px-4 py-2 sm:w-20 md:w-40 text-center text-xl">
+            <Link
+              href="/login"
+              className="text-gray-700 border-2 border-gray-600 rounded-xl px-4 py-2 sm:w-20 md:w-40 text-center text-xl"
+            >
               Login
             </Link>
-            <Link href="/signup" className="bg-gradient-to-r from-customBlue-light via-customBlue-default to-customBlue-dark text-white rounded-xl px-4 py-2 sm:w-20 md:w-40 text-center text-xl">
+            <Link
+              href="/signup"
+              className="bg-gradient-to-r from-customBlue-light via-customBlue-default to-customBlue-dark text-white rounded-xl px-4 py-2 sm:w-20 md:w-40 text-center text-xl"
+            >
               Sign Up
             </Link>
           </div>
@@ -52,6 +74,4 @@ const Navbar = () => {
       </div>
     </nav>
   );
-};
-
-export default Navbar;
+}
