+++ src/middleware.ts
@@ -1,0 +1,11 @@
+import { withAuth } from 'next-auth/middleware';
+
+export default withAuth({
+  pages: {
+    signIn: '/login',
+  },
+});
+
+export const config = {
+  matcher: ['/projects/:path*'],
+};
